+++ packages/stream_video_flutter/lib/src/theme/components/components.dart
@@ -6,3 +6,4 @@ export 'lobby_view_theme.dart';
 export 'participant_label_theme.dart';
 export 'participant_tile_theme.dart';
 export 'picture_in_picture_theme.dart';
+export 'screen_share_selector_theme.dart';